requirement.txt
sqlalchemy
sentence-transformers
chromadb
pydantic-ai
python-dotenv